Codes for rocm-rel-2.0
See the commit in ROCm-OpenCL-Runtime.git:
f6629e2b7e69 ("Codes for rocm-rel-2.0")
Even though this is on roc-2.0.0 branch so
it's ok to just flip the assumption entirely,
we still allow users to specify an earlier
rocm release via the common options infrastructure.
IMHO we should refactor and do in a
backward-compatible way with just one branch.